pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "runpilot"
version = "0.1.0"
description = "RunPilot - a lightweight orchestration tool for reproducible AI training jobs"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "Apache 2.0" }
authors = [
  { name = "Sam Vain" }
]
dependencies = [
  "click>=8.1,<9.0",
  "rich>=13.0,<14.0",
  "httpx>=0.27,<1.0",
  "PyYAML>=6.0,<7.0",
  "requests>=2.32,<3.0",
]

[project.scripts]
runpilot = "runpilot.cli:main"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

